Moving scheduler_factory code from python to C++
