+++ sprokit/src/bindings/python/sprokit/pipeline/scheduler_factory.cxx
@@ -66,33 +66,6 @@ namespace boost
 
 using namespace pybind11;
 
-typedef std::function< object( sprokit::pipeline_t const& pipe,
-                               kwiver::vital::config_block_sptr const& config ) > py_scheduler_factory_func_t;
-
-class python_scheduler_factory
-: public kwiver::vital::plugin_factory
-{
-
-  public:
-
-  python_scheduler_factory( const std::string& type,
-                            const std::string& itype,
-                            py_scheduler_factory_func_t factory );
-
-  virtual ~python_scheduler_factory();
-
-  virtual object create_object(sprokit::pipeline_t const& pipe,
-                               kwiver::vital::config_block_sptr const& config);
-
-private:
-  py_scheduler_factory_func_t m_factory;
-};
-
-
-static object create_py_scheduler( const sprokit::scheduler::type_t&      name,
-                                   const sprokit::pipeline_t&             pipe,
-                                   const kwiver::vital::config_block_sptr config );
-
 static void register_scheduler( sprokit::scheduler::type_t const& type,
                                 sprokit::scheduler::description_t const& desc,
                                 object obj );
@@ -113,7 +86,7 @@ PYBIND11_MODULE(scheduler_factory, m)
       , arg("type"), arg("description"), arg("ctor")
       , "Registers a function which creates a scheduler of the given type.");
 
-  m.def("create_scheduler", &create_py_scheduler
+  m.def("create_scheduler", &sprokit::create_scheduler
       , arg("type"), arg("pipeline"), arg("config") = kwiver::vital::config_block::empty_config()
       , "Creates a new scheduler of the given type.");
 
@@ -155,85 +128,6 @@ class python_scheduler_wrapper
     object const m_obj;
 };
 
-// ------------------------------------------------------------------
-python_scheduler_factory::
-python_scheduler_factory( const std::string& type,
-                          const std::string& itype,
-                          py_scheduler_factory_func_t factory )
-  : plugin_factory( itype )
-  , m_factory( factory )
-{
-  this->add_attribute( CONCRETE_TYPE, type)
-    .add_attribute( PLUGIN_FACTORY_TYPE, typeid(* this ).name() )
-    .add_attribute( PLUGIN_CATEGORY, "scheduler" );
-}
-
-python_scheduler_factory::
-~python_scheduler_factory()
-{ }
-
-object
-python_scheduler_factory::
-create_object(sprokit::pipeline_t const& pipe, kwiver::vital::config_block_sptr const& config)
-{
-  // Call sprokit factory function. Need to use this factory
-  // function approach to handle clusters transparently.
-  return m_factory( pipe, config );
-}
-
-
-// ------------------------------------------------------------------
-object
-create_py_scheduler( const sprokit::scheduler::type_t&      name,
-                     const sprokit::pipeline_t&             pipe,
-                     const kwiver::vital::config_block_sptr config )
-{
-  // First see if there's a C++ scheduler with the name
-  // If that fails, try python instead
-  try
-  {
-    sprokit::scheduler_t c_schd = sprokit::create_scheduler(name, pipe, config);
-    return cast(c_schd);
-  }
-  catch ( sprokit::null_scheduler_registry_config_exception e)
-  {
-    throw sprokit::null_scheduler_registry_config_exception();
-  }
-  catch ( sprokit::null_scheduler_registry_pipeline_exception e)
-  {
-    throw sprokit::null_scheduler_registry_pipeline_exception();
-  }
-  catch ( const std::exception &e) // Now check python
-  {
-    typedef kwiver::vital::implementation_factory_by_name< object > instrumentation_factory;
-    instrumentation_factory ifact;
-
-    kwiver::vital::plugin_factory_handle_t a_fact;
-    try
-    {
-      a_fact = ifact.find_factory( name );
-    }
-    catch ( kwiver::vital::plugin_factory_not_found& e )
-    {
-      auto logger = kwiver::vital::get_logger( "python_scheduler_factory" );
-      LOG_DEBUG( logger, "Plugin factory not found: " << e.what() );
-
-      throw sprokit::no_such_scheduler_type_exception( name );
-    }
-
-    python_scheduler_factory* pf = dynamic_cast< python_scheduler_factory* > ( a_fact.get() );
-    if (0 == pf)
-    {
-      // wrong type of factory returned
-      throw sprokit::no_such_scheduler_type_exception( name );
-    }
-
-    return pf->create_object( pipe, config );
-  }
-
-  return none(); // we shouldn't reach this line
-
-}
 
 
 void
@@ -244,7 +138,7 @@ register_scheduler( sprokit::scheduler::type_t const& type,
   python_scheduler_wrapper const wrap(obj);
 
   kwiver::vital::plugin_manager& vpm = kwiver::vital::plugin_manager::instance();
-  auto fact = vpm.add_factory( new python_scheduler_factory( type,
+  auto fact = vpm.add_factory( new sprokit::python_scheduler_factory( type,
                                                              typeid( object ).name(),
                                                              wrap ) );
