+++ sprokit/src/sprokit/pipeline/scheduler_factory.cxx
@@ -53,6 +53,20 @@ sprokit::scheduler_t create_scheduler( const sprokit::scheduler::type_t&      na
     throw null_scheduler_registry_pipeline_exception();
   }
 
+#ifdef SPROKIT_ENABLE_PYTHON
+  // If python is enabled, we need to check for a python factory first
+  try
+  {
+    auto proc = create_py_scheduler(name, pipe, config);
+    return proc;
+  }
+  catch ( const std::exception &e )
+  {
+    auto logger = kwiver::vital::get_logger( "sprokit.object_factory" );
+    LOG_DEBUG( logger, "No python scheduler found, trying C++");
+  }
+#endif
+
   typedef kwiver::vital::implementation_factory_by_name< sprokit::scheduler > instrumentation_factory;
   instrumentation_factory ifact;
 
@@ -102,5 +116,71 @@ is_scheduler_module_loaded( kwiver::vital::plugin_loader& vpl,
 
   return vpl.is_module_loaded( mod );
 }
+
+// ------------------------------------------------------------------
+// TODO: Most of this code is duplicate, and can be rewritten to use templates
+#ifdef SPROKIT_ENABLE_PYTHON
+python_scheduler_factory::
+python_scheduler_factory( const std::string& type,
+                          const std::string& itype,
+                          py_scheduler_factory_func_t factory )
+  : plugin_factory( itype )
+  , m_factory( factory )
+{
+  this->add_attribute( CONCRETE_TYPE, type)
+    .add_attribute( PLUGIN_FACTORY_TYPE, typeid(* this ).name() )
+    .add_attribute( PLUGIN_CATEGORY, "scheduler" );
+}
+
+python_scheduler_factory::
+~python_scheduler_factory()
+{ }
+
+scheduler_t
+python_scheduler_factory::
+create_object(sprokit::pipeline_t const& pipe, kwiver::vital::config_block_sptr const& config)
+{
+  // Call sprokit factory function. Need to use this factory
+  // function approach to handle clusters transparently.
+  return m_factory( pipe, config ).cast<sprokit::scheduler_t>();;
+}
+
+
+// ------------------------------------------------------------------
+scheduler_t
+create_py_scheduler( const sprokit::scheduler::type_t&      name,
+                     const sprokit::pipeline_t&             pipe,
+                     const kwiver::vital::config_block_sptr config )
+{
+  // First see if there's a C++ scheduler with the name
+  // If that fails, try python instead
+  typedef kwiver::vital::implementation_factory_by_name< pybind11::object > instrumentation_factory;
+  instrumentation_factory ifact;
+
+  kwiver::vital::plugin_factory_handle_t a_fact;
+  try
+  {
+    a_fact = ifact.find_factory( name );
+  }
+  catch ( kwiver::vital::plugin_factory_not_found& e )
+  {
+    auto logger = kwiver::vital::get_logger( "python_scheduler_factory" );
+    LOG_DEBUG( logger, "Plugin factory not found: " << e.what() );
+
+    throw sprokit::no_such_scheduler_type_exception( name );
+  }
+
+  sprokit::python_scheduler_factory* pf = dynamic_cast< sprokit::python_scheduler_factory* > ( a_fact.get() );
+  if (0 == pf)
+  {
+    // wrong type of factory returned
+    throw sprokit::no_such_scheduler_type_exception( name );
+  }
+
+  return pf->create_object( pipe, config );
+}
+
+#endif
+
 
 } // end namespace
